handle float scroll distance in protocol level

+++ app/src/input_events.h
@@ -358,8 +358,8 @@ struct sc_mouse_click_event {
 
 struct sc_mouse_scroll_event {
     struct sc_position position;
-    int32_t hscroll;
-    int32_t vscroll;
+    float hscroll;
+    float vscroll;
     uint8_t buttons_state; // bitwise-OR of sc_mouse_button values
 };
